+++ src/services/workloads-services/make-workloads-base-url.js
@@ -1,0 +1,3 @@
+export const makeWorkloadsBaseUrl = () => {
+  return 'v4/workspace/workloads'
+}
